backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic[email]==2.9.2
python-jose[cryptography]==3.3.0
# Persistence & migrations
SQLAlchemy==2.0.36
alembic==1.13.2
psycopg[binary]==3.2.3
# Testing
pytest==8.3.3
httpx==0.28.1
